Refactored analyser
