+++ src/DotVVM.Analysers/DotVVM.Analysers/DotvvmDiagnosticIds.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DotVVM.Analysers
+{
+    internal static class DotvvmDiagnosticIds
+    {
+        public const string PropertiesSerializabilityRuleId = "DotVVM01";
+    }
+}
